Fix PR comments
1. Rename classes and modules
2. Fix doc
3. Fix tutorial
